+++ snakemake/modules.py
@@ -9,6 +9,8 @@ import re
 from snakemake.exceptions import WorkflowError
 from snakemake.path_modifier import PathModifier
 from snakemake import wrapper
+from snakemake.checkpoints import Checkpoints
+from snakemake.common import Rules, Scatter, Gather
 
 
 def get_name_modifier_func(rules=None, name_modifier=None):
@@ -49,6 +51,7 @@ class ModuleInfo:
         with WorkflowModifier(
             self.workflow,
             config=self.config,
+            base_snakefile=snakefile,
             skip_configfile=self.config is not None,
             skip_validation=self.skip_validation,
             rule_whitelist=self.get_rule_whitelist(rules),
@@ -99,6 +102,7 @@ class WorkflowModifier:
         workflow,
         globals=None,
         config=None,
+        base_snakefile=None,
         skip_configfile=False,
         skip_validation=False,
         rulename_modifier=None,
@@ -110,11 +114,12 @@ class WorkflowModifier:
         namespace=None,
     ):
         self.workflow = workflow
+        self.base_snakefile = base_snakefile
 
         self.globals = (
             globals if globals is not None else dict(workflow.vanilla_globals)
         )
-        if config:
+        if config is not None:
             self.globals["config"] = config
 
         self.skip_configfile = skip_configfile